[MazPicker] Clean code
